backend/pyproject.toml
[tool.poetry]
name = "bes"
version = "0.1.0"
description = "Application for housework sharing."
authors = ["Lucile Lévêque <lululeveq@hotmail.fr>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.2.3"
Pillow = "^10.1.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
djangorestframework-simplejwt = "^5.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
